Add project config and bring the codebase to a clean lint
pyproject.toml sets the ruff rule set, line length and per-file ignores, so
'passes lint' means the same thing locally and in CI. The Databricks notebooks
are excluded from import and undefined-name rules because they execute cell by
cell in a runtime that provides spark, dbutils and display as globals.

Fixed 87 findings. Two suppressions are deliberate and commented: the mojibake
table in dirty.py contains genuinely ambiguous characters, because those are
the actual byte sequences a mis-declared charset produces -- correcting them
would destroy the defect the table exists to reproduce.

+++ src/vivo360/config.py
@@ -238,5 +238,5 @@ def planned_fact_rows(profile: ScaleProfile) -> dict[str, int]:
     """Full planned row count for every fact table in the model."""
     plan = dict(profile.facts)
     for name, base in SECONDARY_FACTS.items():
-        plan[name] = max(1, int(round(base * profile.secondary_scale)))
+        plan[name] = max(1, round(base * profile.secondary_scale))
     return plan
